+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Implemented enhancements
 - Move form choices to [Constants][directory-constants].
+- [TT-1271](https://uktrade.atlassian.net/browse/TT-1271) Allow free-text entry of "other" products and services.
 
 ### Fixed bugs:
 - Upgraded urllib3 to fix [vulnerability](https://nvd.nist.gov/vuln/detail/CVE-2019-11324)